+++ sdk/postgresql/Azure.ResourceManager.PostgreSql/tests/Scenario/PostgreSqlFlexibleServerTests.cs
@@ -387,25 +387,26 @@ namespace Azure.ResourceManager.PostgreSql.Tests
         public async Task Replica(bool vnetEnabled)
         {
             var sourceServerName = Recording.GenerateAssetName("pgflexserver");
-            var replicaServerName = new string[2];
+            var replicaServerName = new string[3];
             var vnetName = Recording.GenerateAssetName("vnet");
             var sourceSubnetName = Recording.GenerateAssetName("subnet");
-            var replicaSubnetName = new string[2];
+            var replicaSubnetName = new string[3];
+            var virtualEndpointName = Recording.GenerateAssetName("vendpoint");
 
             var rg = await CreateResourceGroupAsync(Subscription, "pgflexrg", AzureLocation.EastUS);
             var serverCollection = rg.GetPostgreSqlFlexibleServers();
 
-            var replicaSubnetID = new ResourceIdentifier[2];
+            var replicaSubnetID = new ResourceIdentifier[3];
             ResourceIdentifier vnetID;
             ResourceIdentifier subnetID = null;
             PrivateDnsZoneResource sourcePrivateDnsZone = null;
 
-            var replicaSubnet = new SubnetResource[2];
-            var replicaPrivateDnsZone = new PrivateDnsZoneResource[2];
+            var replicaSubnet = new SubnetResource[3];
+            var replicaPrivateDnsZone = new PrivateDnsZoneResource[3];
 
-            for (var i = 0; i < 2; i++)
+            for (var i = 0; i < 3; i++)
             {
-                replicaServerName[i] = Recording.GenerateAssetName("pgflexserver");
+                replicaServerName[i] = Recording.GenerateAssetName("pgflexserverrep");
                 replicaSubnetName[i] = Recording.GenerateAssetName("subnet");
             }
 
@@ -437,7 +438,7 @@ namespace Azure.ResourceManager.PostgreSql.Tests
                 //SubnetResource subnetResource = (await subnetCollection.CreateOrUpdateAsync(WaitUntil.Completed, subnetName2, subnetData)).Value;
                 vnetID = vnetResource.Data.Id;
                 subnetID = vnetResource.Data.Subnets[0].Id;
-                for (var i = 0; i < 2; ++i)
+                for (var i = 0; i < 3; ++i)
                 {
                     var replicaSubnetOperation = await vnetResource.GetSubnets().CreateOrUpdateAsync(WaitUntil.Completed, replicaSubnetName[i], new SubnetData()
                     {
@@ -457,7 +458,7 @@ namespace Azure.ResourceManager.PostgreSql.Tests
                     replicaSubnetID[i] = replicaSubnetOperation.Value.Data.Id;
                 }
                 sourcePrivateDnsZone = await CreatePrivateDnsZone(sourceServerName, vnetID, rg.Data.Name);
-                for (var i = 0; i < 2; ++i)
+                for (var i = 0; i < 3; ++i)
                 {
                     replicaPrivateDnsZone[i] = await CreatePrivateDnsZone(replicaServerName[i], vnetID, rg.Data.Name);
                 }
@@ -568,11 +569,89 @@ namespace Azure.ResourceManager.PostgreSql.Tests
             Assert.AreEqual(PostgreSqlFlexibleServerReplicationRole.Primary, replica0Server.Data.ReplicationRole);
             Assert.IsNull(replica0Server.Data.SourceServerResourceId);
 
-            // Can't delete source server if it has replicas
+            // Create replica 2
+            var replica2ServerData = new PostgreSqlFlexibleServerData(rg.Data.Location)
+            {
+                SourceServerResourceId = sourceServer.Id,
+                AvailabilityZone = "2",
+                CreateMode = PostgreSqlFlexibleServerCreateMode.Replica,
+            };
+            if (vnetEnabled)
+            {
+                replica2ServerData.Network = new PostgreSqlFlexibleServerNetwork()
+                {
+                    DelegatedSubnetResourceId = replicaSubnetID[2],
+                    PrivateDnsZoneArmResourceId = replicaPrivateDnsZone[2].Id,
+                };
+            }
+            var replica2ServerOperation = await serverCollection.CreateOrUpdateAsync(WaitUntil.Completed, replicaServerName[2], replica2ServerData);
+            var replica2Server = replica2ServerOperation.Value;
+
+            Assert.AreEqual(replicaServerName[2], replica2Server.Data.Name);
+            Assert.AreEqual("2", replica2Server.Data.AvailabilityZone);
+            Assert.AreEqual(sourceServer.Data.Sku.Name, replica2Server.Data.Sku.Name);
+            Assert.AreEqual(sourceServer.Data.Sku.Tier, replica2Server.Data.Sku.Tier);
+            Assert.AreEqual(PostgreSqlFlexibleServerReplicationRole.AsyncReplica, replica2Server.Data.ReplicationRole);
+            Assert.AreEqual(sourceServer.Id, replica2Server.Data.SourceServerResourceId);
+            Assert.AreEqual(0, replica2Server.Data.ReplicaCapacity);
+            if (vnetEnabled)
+            {
+                Assert.AreEqual(replicaSubnetID[2], replica2Server.Data.Network.DelegatedSubnetResourceId);
+                Assert.AreEqual(replicaPrivateDnsZone[2].Id, replica2Server.Data.Network.PrivateDnsZoneArmResourceId);
+            }
+
+            replicaList = await serverCollection.GetReplicasAsync(sourceServerName).ToEnumerableAsync();
+            Assert.AreEqual(2, replicaList.Count);
+
+            // Create Virtual Endpoint
+            VirtualEndpointResourceCollection virtualEndpointSourceCollection = sourceServer.GetVirtualEndpointResources();
+            VirtualEndpointResourceData virtualEndpointData = new VirtualEndpointResourceData()
+            {
+                EndpointType = VirtualEndpointType.ReadWrite,
+                Members = { sourceServerName },
+            };
+            VirtualEndpointResource virtualEndpointResource = (await virtualEndpointSourceCollection.CreateOrUpdateAsync(WaitUntil.Completed, virtualEndpointName, virtualEndpointData)).Value;
+            var virtualEndpointDataName = virtualEndpointResource.Data.Name;
+            virtualEndpointSourceCollection = sourceServer.GetVirtualEndpointResources();
+            virtualEndpointResource = (await virtualEndpointSourceCollection.GetAsync(virtualEndpointName)).Value;
+            Assert.AreEqual(virtualEndpointDataName, virtualEndpointResource.Data.Name);
+
+            // Update Virtual Endpoint
+            VirtualEndpointResourceCollection virtualEndpointUpdateCollection = sourceServer.GetVirtualEndpointResources();
+            virtualEndpointData = new VirtualEndpointResourceData()
+            {
+                EndpointType = VirtualEndpointType.ReadWrite,
+                Members = { replicaServerName[2] },
+            };
+            VirtualEndpointResource virtualEndpointUpdateResource = (await virtualEndpointUpdateCollection.CreateOrUpdateAsync(WaitUntil.Completed, virtualEndpointName, virtualEndpointData)).Value;
+            virtualEndpointDataName = virtualEndpointUpdateResource.Data.Name;
+            virtualEndpointUpdateCollection = sourceServer.GetVirtualEndpointResources();
+            virtualEndpointResource = (await virtualEndpointUpdateCollection.GetAsync(virtualEndpointName)).Value;
+            Assert.AreEqual(virtualEndpointDataName, virtualEndpointResource.Data.Name);
+
+            // Perform switchover on replica 2
+            var replica2ServerUpdate = await replica2Server.UpdateAsync(WaitUntil.Completed, new PostgreSqlFlexibleServerPatch()
+            {
+                Replica = new Replica()
+                {
+                    PromoteMode = ReadReplicaPromoteMode.Switchover,
+                    PromoteOption = ReplicationPromoteOption.Forced,
+                },
+            });
+            replica2Server = replica2ServerUpdate.Value;
+
+            Assert.AreEqual(replicaServerName[2], replica2Server.Data.Name);
+            Assert.AreEqual(PostgreSqlFlexibleServerReplicationRole.Primary, replica2Server.Data.ReplicationRole);
+            Assert.IsNull(replica2Server.Data.SourceServerResourceId);
+
+            // Delete Section
+            await virtualEndpointResource.DeleteAsync(WaitUntil.Completed);
+
+            // Can't delete source server (now replica 2) if it has replicas
             RequestFailedException deleteException = null;
             try
             {
-                await sourceServer.DeleteAsync(WaitUntil.Completed);
+                await replica2Server.DeleteAsync(WaitUntil.Completed);
             }
             catch (RequestFailedException ex)
             {
@@ -583,8 +662,11 @@ namespace Azure.ResourceManager.PostgreSql.Tests
             // Delete replica 1
             await replica1Server.DeleteAsync(WaitUntil.Completed);
 
-            // Now delete source server
+            // Delete original source server
             await sourceServer.DeleteAsync(WaitUntil.Completed);
+
+            // Now delete the new source server
+            await replica2Server.DeleteAsync(WaitUntil.Completed);
         }
 
         [TestCase("11", "12")]
@@ -641,7 +723,7 @@ namespace Azure.ResourceManager.PostgreSql.Tests
         {
             var serverName = Recording.GenerateAssetName("pgflexserver");
 
-            var rg = await CreateResourceGroupAsync(Subscription, "pgflexrg", AzureLocation.EastUS);
+            var rg = await CreateResourceGroupAsync(Subscription, "pgflexrg", AzureLocation.SouthCentralUS);
             var serverCollection = rg.GetPostgreSqlFlexibleServers();
 
             var serverOperation = await serverCollection.CreateOrUpdateAsync(WaitUntil.Completed, serverName, new PostgreSqlFlexibleServerData(rg.Data.Location)
@@ -674,6 +756,16 @@ namespace Azure.ResourceManager.PostgreSql.Tests
             Assert.AreEqual(automaticBackup.Data.Name, backups[0].Data.Name);
             Assert.AreEqual(automaticBackup.Data.BackupType, backups[0].Data.BackupType);
             Assert.AreEqual(automaticBackup.Data.CompletedOn, backups[0].Data.CompletedOn);
+
+            // on demand backup section
+            PostgreSqlFlexibleServerBackupCollection collection = server.GetPostgreSqlFlexibleServerBackups();
+            string onDemandBackupName = "backup_test_20241025159";
+            var onDemanBackup = (await collection.CreateOrUpdateAsync(WaitUntil.Completed, onDemandBackupName)).Value;
+            PostgreSqlFlexibleServerBackupData onDemanBackupData = onDemanBackup.Data;
+
+            collection = server.GetPostgreSqlFlexibleServerBackups();
+            bool result = await collection.ExistsAsync(onDemanBackupData.Name);
+            Assert.IsTrue(result);
         }
 
         [TestCase]